Update construct-your-world.md (#148)
## Changes

<!-- Please summarize your changes. -->
One of the images toward the end had 3A instead of 3a, effectively
breaking the image. From the bottom of my heart: my bad.
<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [ ] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [ ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [ ] To the best of my knowledge, all proposed changes are accurate.