security-tactics improvements and additions (#8)
## Changes

<!-- Please summarize your changes. -->
I made the following changes to improve pre-existing code:
- The `typeValidator` tuple type validator in the Remote Runtime Type
Validation was created on-the-fly every single time a request was made
from the client. If something can be made once and reused infinitely, it
should be!
- Instead of throwing an error on the server when the `typeValidator`
type check fails, which could trigger analytics requests and other heavy
error handling logic (among other things) on an event that is *not* rate
limited, it silently fails and returns that information to the client to
receive, display and act on accordingly. In this example it isn't a
fatal error that should be logged in that way.
- Adjusted some of the "highlight" data in the Lua code blocks. I
presume these are used to highlight important parts of the code samples,
though I've never seen this myself on the creator docs. Regardless, they
should be more accurate for the code snippets now.

I added the following additions to the article:
- Added an example of "instance spoofing" under Data Validation that
ties in with the existing "In-Experience Shop" example. This is an
incredibly common and easily overlooked exploit that can be taken
advantage of in many games, and definitely deserves a spot in this
article.
- Added an in-depth list of various attacks that can prevent DataStore
saving from malicious client input under the right circumstances, and
how to prevent them. The listed attacks can be detrimental to games and
could result in item duplication among other game-breaking actions based
on the type of game.

I've given everything a good look over based on the style guide and
rewritten a lot from my initial draft. Please correct anything that I
might've missed. Cheers!

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.