Improve code and instructions in selling-items.md (#166)
## Changes

Added if statements to check if certain instances were valid. Fixed
section where the script would error if a humanoid with no player (an
NPC) touched the platform. Changed "playerGold.Value = playerGold.Value
+ totalSell" to "playerGold.Value += totalSell" to better reflect the
code.

These errors are in other docs that use :FindFirstChild and do not check
if the returned instance isn't nil.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.