Improve code in cleanup-and-reset.md (#195)
## Changes

Switched variable placements for easier readability. Added variables to
store objects that `:FindFirstChild()` returned for readability and
performance. Switched pairs to generalized iteration for readability and
performance.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------