Update creating-a-timed-bridge.md (#204)
The original documentation requires the player to change the CanvasSize,
TextScaled and Text properties in SurfaceGui properties while these
three properties belong to TextLabel.

## Changes

<!-- Please summarize your changes. -->
The original documentation instructs the player to modify the
CanvasSize, TextScaled, and Text properties in the SurfaceGui
properties. However, these properties actually belong to the TextLabel
component, not the SurfaceGui.

The proposed changes rectify this issue by providing the accurate
location of the properties to be modified, which is the TextLabel
properties.


<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------