Expand and correct documentation on data store string sanitisation. (#203)
The documentation that explains how to sanitise strings is problematic
and does not properly solve the problem it claims it solves. Data stores
can save valid UTF-8 while the provided fix which was borrowed from my
[github
gist](https://gist.github.com/TheGreatSageEqualToHeaven/e0e1dc2698307c93f6013b9825705899#patch-1)
only allows ascii characters to be passed, it is not a real fix and
should not be recommended.

This documentation should also be further updated if my [feature
request](https://devforum.roblox.com/t/datastoreservice-method-to-verify-if-a-value-is-saveable-bring-more-attention-to-malicious-strings/2658551)
were to be accepted and introduce a proper solution.


![image](https://github.com/Roblox/creator-docs/assets/85808999/e9427833-b710-45a0-9d69-034d25304857)

Edit:

Luau now has [stricter UTF-8
validation](https://github.com/luau-lang/rfcs/blob/master/docs/behavior-stricter-utf8-library.md)
as of 2 weeks ago so we can replace the UTF-8 verification function with
`utf8.len` and expect it to work properly.