CLI-88769 - Add API documentation for UnreliableRemoteEvents (#331)
## Changes

Add V1 description of UnreliableRemoteEvents. Note that much of this is
the same as RemoteEvents and was ported over with minor changes. This is
because this new instance is a variant of the RemoteEvent instance.

Essentially what I did was added how this instance differs from
RemoteEvents + added the caveats that developers should be aware of when
using UREs.

https://devforum.roblox.com/t/introducing-unreliableremoteevents/2724155

## Checks

By submitting your pull request for review, you agree to the following:

- [ ] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [ ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [ ] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>