Update PreloadAsync in ContentProvider.yaml (#364)
## Changes

<!-- Please summarize your changes. -->
Updated long outdated description of PreloadAsync method to reflect
actual behavior in studio and in game. The method supports passing in a
list of ContentId strings, but documentation is unclear about this.
Description of parameters also introduces confusion with the mismatch in
parameter name and the rest of the method description only describing
instances. ContentId strings passed into the method must also only be
image assets, a change made in 2017 but not mentioned in the docs.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

https://devforum.roblox.com/t/contentproviderpreloadasync-has-an-extremely-high-failure-rate-when-using-content-strings/2660912

https://devforum.roblox.com/t/contentproviderpreloadasync-does-not-preload-image-or-texture-assets-in-either-assetid-form-or-instance-form-appears-to-be-a-regression/1637450/25
## Checks

By submitting your pull request for review, you agree to the following:

- [ x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [ x] To the best of my knowledge, all proposed changes are accurate.