Add markdown escape to TextChatService.yaml (#382)
## Changes

The last pull request for this problem fixed most of the problem, but it
looks like the bar character in `Class.TextChannel | TextChannels` is
now being used for the markdown table.


![IncorrectMarkdown](https://github.com/Roblox/creator-docs/assets/19847497/a8ea0815-7bd3-4e92-bea5-4e6a13067a7f)

This is the reason why this pull request adds an escape character before
the bar. It is unknown if this will break the automatic link:

`Class.TextChannel \| TextChannels`

I tried to see what was done in such a scenario on other pages, but I
was unable to find a similar scenario. The closest I found was this
table on
[Datatype.Vector3](https://create.roblox.com/docs/reference/engine/datatypes/Vector3)
However, it turned out that table just used pure html tags for its table
and it doesn't override the default text link.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------