Fix broken backticks in in-line code samples (#452)
## Changes
A few pages have broken backtick syntax for in-line code samples. For
example, they begin and end with four ` instead of three. Additionally,
before the closing backticks, there is a duplicate set of three
backticks followed by a period or quote. This duplicate set leaks
markdown syntax into the code sample, as shown in this screenshot at the
bottom:

![image](https://github.com/Roblox/creator-docs/assets/55513323/e010c338-5c24-4899-8476-48739bec8800)

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.