Fix incorrect endpoint in External Catalog Queries (#467)
## Changes

* Changed `catalog.roblox.com/v2/search/items/details` to
`catalog.roblox.com/v1/search/items/details`. The current documented
parameters are for V1 and not V2. (i.e. `category` parameter in V2
exists as `categoryFilter` and `subcategory` does not exist in V2).
* Fixed example URL for searching Gears

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.