Fix code sample for variadics (#538)
## Changes

This PR fixes a code sample that uses the wrong string function to
convert a series of numbers representing bytes into a string. The
current, incorrect, version calls `string.byte` which does not return
anything when called with only numbers. The fixed version calls
`string.char` which correctly returns "secret" here as indicated by the
original comment.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.