Update raycasting.md (#564)
## Changes

Edited warning. Present warning is a bit misleading. Seems to imply that
there is a link between the origin and the magnitude of the ray. I feel
like this could potentially confuse beginners into thinking the second
param is a destination.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.