Fix double spacing in prints (#588)
## Changes

<!-- Please summarize your changes. -->
`print(value1, value2)` already puts a space in between the values, so
`print(value1.." ", value2)` puts two spaces. This PR removes the extra
space.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.