Clarify 2D classic clothing support (#666)
## Changes

<!-- Please summarize your changes. -->
Clarify that user generated avatars which do not come from the
Marketplace can support 2D clothing. They can be UV mapped to fit the
offical Roblox clothing schema, however not offically as there is no
documentation for it

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->
Examples of rigs that support 2D clothing which do not come from the
Marketplace:
https://devforum.roblox.com/t/dogu15-rig-layered/532832
https://devforum.roblox.com/t/marys-custom-r15-character-model/2828073
## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.