Update Secret.yaml (#680)
## Changes

<!-- Please summarize your changes. -->
- Slightly altered description for AddSuffix to provide a clearer
usecase scenario in adding additional query parameters to the end of an
authenticated url using secrets.

- Changed example code for AddSuffix to work as intended with AddSuffix
instead of AddPrefix. Previous code appended a url to the end of a
secret which is not very useful. Provided code prefixes a url to a
secret and then suffixes further query parameters to the url+secret
combination.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------