Remove physics point from designing.md (#697)
## Changes

The statement about anchoring and collision properties on parts isn't
totally accurate. There are no "unnecessary" physics calculations. Parts
that are not being interacted with will fall asleep, and not do any
work. Same goes for the notes about collision.

The "Improving" page already covers this same information, with more
detail, and predicated with how a developer could determine when they
need to optimize simulation. This approach is much preferred when it
comes to physics simulation.
## Checks

By submitting your pull request for review, you agree to the following:

- [ ] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [ ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [ ] To the best of my knowledge, all proposed changes are accurate.