Minor typo in data-store-api-handling.md (#740)
## Changes

<!-- Please summarize your changes. -->

The original alert info for the Input Validation section said
"Perfect-Encoding", while the linked header is "Percent-Encoding". The
commit changes "Perfect" into "Percent".

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.