Update index.md (#750)
Replaced 'ServerScriptStorage' by 'ServerScriptService' to match the
folder's actual name.

## Changes

<!-- Please summarize your changes. -->
Name of folder in documentation text was different from actual name in
'Intro to Studio - Catapult & Platforms' model. I corrected that
mismatch in the text.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [ ] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [ ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [ ] To the best of my knowledge, all proposed changes are accurate.