Fix inaccuracies in spawn and defer descriptions. (#892)
## Changes

- Fixes `task.defer` description that stated that it runs the code on
the next resumption cycle. This is false. It runs at the end of the
_current_ resumption cycle.
- Fixes `task.defer` and `task.spawn` inaccuracy that states they do not
return anything. They both return a thread object, which is reflected in
the return value section of the documentation. Thus, these sentences are
removed.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------