Remove use of the word "force" in describing ApplyImpulse methods (#976)
## Changes

Since the various ApplyImpulse functions apply a "pure" impulse
(resulting in an instantaneous change in velocity), using the word
"force" is misleading. A force must be applied over some finite time
span to produce a velocity change, whereas an impulse is applied
instantaneously. This has led to confusion among developers.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.