General Iteration #2 (#974)
## Changes

Some pages have pairs and ipairs creeping back in or that were missed in
my previous PR for general iteration. This PR changes them to general
iteration. Will update other pages in the future to further remove
references to pairs and ipairs however that work is out of scope for
this PR as it will be a larger change to the docs.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>