Change www.authorize.roblox.com to authorize.roblox.com (#1030)
## Changes

This pull request changes the link from https://www.authorize.roblox.com
to the correct URL which is https://authorize.roblox.com

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.