Update AnimationTrack.yaml (#1027)
## Changes

<!-- Please summarize your changes. -->
animation controller methods are deprecated in favor of directly using
animator

there's some code samples I want to change but I don't know how to.
Where can I find it? "Please note that AnimationPlayed is a member of
both Humanoid and AnimationController which this example would also work
on." <- i cant find this line but i see it on the docs page
<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [ x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x ] To the best of my knowledge, all proposed changes are accurate.