Remove outdated bullet point (#1033)
## Changes

Removes an outdated bullet point regarding the use of
`task.wait/task.spawn/task.delay`. It seems this is a holdover from the
older global functions `wait/spawn/delay`, and that someone had simply
added the `task` prefix to those older functions in this paragraph (you
can see this was forgotten on the last reference to `spawn()`).

To the best of my understanding, the newer task library has no
throttling issues like the older global functions, and thus the docs
should not be recommending against their usage.

Relevant links:
- https://devforum.roblox.com/t/task-library-now-available/1387845
- https://create.roblox.com/docs/reference/engine/libraries/task

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.