Update MarketplaceService.yaml (#1035)
## Changes

Added/removed/moved a ton of things for GetProductInfo():
- Moved product id to product section.
- Moved IconImageAssetId to asset section.
- Removed gamepass/dev products section (these fields apply to all
products and assets not just gamepasses/dev products).
- Added seperate table for collectibles information with the appropriate
fields.
- Moved sale location information to seperate table with more
information.
- Fixed typos where "times" was instead typed as "items".
- Added TargetId to product section which wasn't present for some
reason.

With all this the page should now be consistent with the result of
[MarketplaceService:GetProductInfo()](https://create.roblox.com/docs/reference/engine/classes/MarketplaceService#GetProductInfo):


![image](https://github.com/user-attachments/assets/18b08a71-aa98-4e1c-821a-39a6958cdc29)

Credit to [@Valkenheim](https://devforum.roblox.com/u/Valkenheim) for a
[devforum
post](https://devforum.roblox.com/t/collectiblesitemdetails-not-covered-in-documentation/3505808)
regarding the inconsistencies on this page, but this page was kind of a
mess all around.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>