Update TextChatCommand.yaml (#1046)
A documentation report was filed by a dev regarding the clarity of using
TextChatCommands. Proposed solution is to directly link the examples
page to the documentation guide to improve visibility to developers.

## Changes

<!-- Please summarize your changes. -->
I've added the link for custom TextChatCommand usage to the
TextChatCommands documentation page.
<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->
https://roblox.atlassian.net/browse/CREATORBUG-11855
## Checks

By submitting your pull request for review, you agree to the following:

- [ ] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [ ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [ ] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: julia <jbrowne@roblox.com>