Added code snippet for deriving TextChatMessageProperties (#1048)
## Changes

I was trying to figure out how to derive TextChatMessageProperites, and
I couldn't find it anywhere. This PR aims to add a code snippet to the
TextChatMessageProperties page (
https://create.roblox.com/docs/reference/engine/classes/TextChatMessageProperties
) that will show how to do so.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------