Replace their with there and mention that Lighting.Technology cannot be accessed from Roblox Studio (#1057)
## Changes

Replaced "their are plans in the future" with "there are plans in the
future" in the How to use Scale section. Also I fixed a misinformation
stating that Lighting.Technology can be changed from Roblox Studio - it
can't be even read from studio.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------